requirements.txt
pyzmq
pytest
pytest-asyncio
pytest-timeout
gino
aiohttp==2.3.10
aiohttp-jinja2
aiosocks==0.2.6
aiopg
peewee-async
fake_useragent
lxml
py_mini_racer
bs4
